Administrative changes

+++ LC3_Web/rl_universal.php
@@ -6,6 +6,8 @@ include("./incl/header.php");
 if (!isset($_SESSION["client"])) $_SESSION["client"] = 1;
 if ($_POST["client"]) $_SESSION["client"] = $_POST["client"];
 
+$ia_8 = is_admin(8); 
+
 
 $clients = array();
 foreach($dbh->query("SELECT [КодКлиента] AS clientid, [НаименованиеКлиента] AS name, [ПолноеНаименование] AS fullname FROM [dbo].[_CL_Clients]") as $row) {
@@ -21,12 +23,14 @@ $message = "";
 
 $aa["CORG"] = $_POST["CORG"]; $aa["rtype"] = $_POST["rtype"];
 $aa["cbtypes"] = $_POST["cbtypes"]; $aa["ltype"] = $_POST["ltype"]; 
-$aa["minval"] = $_POST["minval"]; $aa["maxval"] = $_POST["maxval"];
+$aa["minval"] = $_POST["minval"]; $aa["maxval"] = $_POST["maxval"]; $aa["nobuy"] = $_POST["nobuy"];
 $aa["cgrid"] = $_POST["corg_list"]; $aa["issueid"] = $_POST["issues_list"];
 $aa["fl_cgrt"] = $_POST["fl_cgrt"]; $aa["fl_cgid"] = $_POST["fl_cgid"];  $aa["fl_tliss"] = $_POST["fl_tliss"];
 
 
-if ( ($_POST["fl"] == "del") && (isset($_POST["ed_rid"])) ) {
+
+
+if ( ($ia_8) && ($_POST["fl"] == "del") && (isset($_POST["ed_rid"])) ) {
 	$aa["fl"] = 1;
 	$query = "DELETE FROM RL_Universal WHERE id = '".$_POST["ed_rid"]."'";
 	$dbh->query($query);
@@ -34,16 +38,16 @@ if ( ($_POST["fl"] == "del") && (isset($_POST["ed_rid"])) ) {
 }
 
 //	Вход в редактирование записи
-if ($_POST["fl"] == "edit") {
+if ( ($ia_8) && ($_POST["fl"] == "edit")) {
 	if (isset($_SESSION["client"]) && ($_SESSION["client"] > 0)) {
 		$aa["fl"] = 1; $aa["mode"] = "edit"; $aa["ed_rid"] = $_POST["ed_rid"];
 	
-		$query = "SELECT CORG, CORGid, RestrictType, IssueRid, TypesList, LimitType, Min, Max 
+		$query = "SELECT CORG, CORGid, RestrictType, IssueRid, TypesList, LimitType, Min, Max, NoBuy
 						FROM RL_Universal WHERE id = '".$aa["ed_rid"]."' AND ClientId = '".$_SESSION["client"]."'";
 		foreach($dbh->query($query) as $row) {
 			$aa["CORG"] = $row["CORG"]; $aa["rtype"] = $row["RestrictType"];
 			$aa["cbtypes"] = explode(";", $row["TypesList"]); $aa["ltype"] = $row["LimitType"]; 
-			$aa["minval"] = $row["Min"]; $aa["maxval"] = $row["Max"];
+			$aa["minval"] = $row["Min"]; $aa["maxval"] = $row["Max"]; $aa["nobuy"] = $row["NoBuy"];
 			$aa["cgrid"] = $row["CORGid"]; $aa["issueid"] = $row["IssueRid"];
 		}
 	} else {
@@ -51,11 +55,14 @@ if ($_POST["fl"] == "edit") {
 	}	
 }
 
+		//print_r($_POST);
+
 //	Добавление или редактирование записи
-if ($_POST["fl"] == "add") {
+if ( ($ia_8) && ($_POST["fl"] == "add")) {
 	if (isset($_SESSION["client"]) && ($_SESSION["client"] > 0)) {	
 		$cbt = ""; foreach ($aa["cbtypes"] as $t) $cbt .= $t.";";
 		$aa["fl"] = 1;
+
 		if (!isset($_POST["corg_list"]) || (($_POST["rtype"] == 'I') && !isset($_POST["issues_list"])) ||
 			(($_POST["rtype"] != 'I') && ($cbt == "")) || !isset($_POST["minval"]) || !isset($_POST["minval"])) {
 				$message .= "Не заданы все обязательные параметры (эмитент или группа, список типов бумаг или выпуск, размер ограничений)";
@@ -64,7 +71,8 @@ if ($_POST["fl"] == "add") {
 						", @RestrictType = '".$_POST["rtype"]."', @IssueRid = ".($_POST["rtype"] == 'I' ? "'".$_POST["issues_list"]."'" : 'NULL').
 						", @TypesList = ".($_POST["rtype"] == 'I' ? "NULL" : "'".$cbt."'").", @LimitType = '".$_POST["ltype"]."'".
 						", @Min = ".$_POST["minval"].", @Max = ".$_POST["maxval"].
-						( (isset($_POST["ed_rid"]) && ($_POST["ed_rid"] != "")) ? ", @Ridtodel = '".$_POST["ed_rid"]."'" : "");
+						", @NoBuy = ".((isset($_POST["nobuy"]) && ($_POST["nobuy"] == 1)) ? "1" : "0").
+						( (isset($_POST["ed_rid"]) && ($_POST["ed_rid"] != "")) ? ", @Ridtodel = '".$_POST["ed_rid"]."'" : "");					
 				$dbh->query($query);
 				$message .= "Ограничение сохранено";					
 			}
@@ -74,7 +82,7 @@ if ($_POST["fl"] == "add") {
 }
 
 //	Массовое копирование
-if ($_POST["fl"] == "copyall") {
+if ( ($ia_8) && ($_POST["fl"] == "copyall")) {
 	$aa["fl"] = 1;
 	$aa["cb"] = $_POST["cb"];
 	foreach ($aa["cb"] as $k => $v) {
@@ -85,7 +93,7 @@ if ($_POST["fl"] == "copyall") {
 }
 
 //	Удаление отмеченных
-if ($_POST["fl"] == "delch") {
+if ( ($ia_8) && ($_POST["fl"] == "delch")) {
 	$aa["fl"] = 1;
 	$aa["cb"] = $_POST["cb"];
 	foreach ($aa["cb"] as $k => $v) {
@@ -96,7 +104,7 @@ if ($_POST["fl"] == "delch") {
 }
 
 //	Получаем названия для контрагентов или групп и выпусков
-if ($aa["fl"] == 1) {
+if ( ($ia_8) && ($aa["fl"] == 1)) {
 	if ($aa["CORG"] == "G") { 
 		$query = "SELECT [Rid] AS [rid], [GroupName] AS [name] FROM [dbo].[CL_Groups] WHERE [Rid] = '".$aa["cgrid"]."'";
 	} else {  
@@ -113,7 +121,7 @@ if ($aa["fl"] == 1) {
 
 $query = "SELECT  rl.[id], rl.[ClientId], rl.[CORG], rl.[CORGid], CASE WHEN rl.[CORG] = 'C' THEN c.name ELSE g.GroupName END AS CORGname
 			,rl.[RestrictType], rl.[IssueRid], CASE WHEN rl.RestrictType = 'I' THEN i.FullName ELSE NULL END AS Issue
-			,rl.[TypesList], rl.[LimitType], rl.[Min], rl.[Max]
+			,rl.[TypesList], rl.[LimitType], rl.[Min], rl.[Max], rl.[NoBuy]
 			FROM [RL_Universal] AS rl LEFT JOIN [_CL_Contragents] AS c ON (c.rid = rl.CORGid)
 			LEFT JOIN [CL_Groups] AS g ON (g.Rid = rl.CORGid) LEFT JOIN [_CL_URL_issues] AS i ON (i.rid = rl.IssueRid) WHERE rl.[ClientId] = ".$_SESSION["client"].
 			" ORDER BY rl.[CORG], rl.[RestrictType], CORGname, rl.[TypesList], Issue";
@@ -131,6 +139,7 @@ $smarty->assign("cbtypes", $cbtypes);
 $smarty->assign("message", $message);
 $smarty->assign("aa", $aa);
 $smarty->assign("rlu", $rlu);
+$smarty->assign("ia_8", $ia_8);
 
 $smarty->display("templates/rl_universal.tpl");
